Update task1_code.cpp
changed int to ll for no overflow

+++ tasks/task_1/task1_code.cpp
@@ -39,7 +39,7 @@ int main()
     #endif
 
     ios_base::sync_with_stdio(false);
-    int i,w,h,u1,d1,d2,dm,dmn,pre[100001];
+    ll i,w,h,u1,d1,d2,dm,dmn,pre[100001];
     for(i=1;i<100001;i++)
     {
         pre[i] += (pre[i - 1] + i);
